+++ test/format/E.js
@@ -5,7 +5,7 @@ var format = koyomi.format.bind(koyomi);
 var eq = require('assert').equal;
 
 // 営業日・休業日
-// EE Ew E
+// Ew E
 eq(format('2015-1-1', 'Ew'), '休業日');
 eq(format('2015-1-1', 'E' ), '休業日');