+++ test/format/combi.js
@@ -6,4 +6,4 @@ var eq = require('assert').equal;
 
 // 混合
 eq(format('2015-4-10', 'YMMDD'), '20150410');
-eq(format('2015-4-10', 'GGGggg'), '平成H平成H');
+eq(format('2015-4-10', 'GGGggg'), '2675平成H');