+++ test/format/custom.js
@@ -1,0 +1,11 @@
+// フォーマット
+var Koyomi = require('../..');
+var format = Koyomi.format.bind(Koyomi);
+var eq = require('assert').equal;
+
+// WAREKI wareki
+eq(format('2015-4-10', 'WAREKI'), '平成二十七年四月十日');
+eq(format('2015-4-10', 'wareki'), '平成二十七年四月十日');
+
+eq(format('1989-1-7', 'WAREKI'), '平成元年一月七日');
+eq(format('1989-1-7', 'wareki'), '昭和六十四年一月七日');